Fixed Web App on Mobile Not Working
It was the font size all along... The font size was preventing the buttons from getting smaller... It was working on desktop because it seems like the pixel count is different therefore the font size is different. I fixed it by just forcing the buttons to expand to its parent's size then constrained the parent. For some reason this causes the text to reflow instead of just blocking the minimization of the buttons.